refactor: 인덱스에서 내보내기
