+++ src/components/Product/index.ts
@@ -4,3 +4,4 @@ export { default as ProductList } from './ProductList/ProductList';
 export { default as ProductOverviewItem } from './ProductOverviewItem/ProductOverviewItem';
 export { default as ProductRecipeList } from './ProductRecipeList/ProductRecipeList';
 export { default as ProductTitle } from './ProductTitle/ProductTitle';
+export { default as ProductPreviewList } from './ProductPreviewList/ProductPreviewList';
