+++ src/hooks/queries/product/index.ts
@@ -1,5 +1,6 @@
 export { useCategoryFoodQuery } from './useCategoryQuery';
 export { useCategoryStoreQuery } from './useCategoryQuery';
+export { useCategoryQuery } from './useCategoryQuery';
 export { default as useInfiniteProductsQuery } from './useInfiniteProductsQuery';
 export { default as useInfiniteProductReviewsQuery } from './useInfiniteProductReviewsQuery';
 export { default as useProductDetailQuery } from './useProductDetailQuery';
